requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9
supabase>=2.7.0
chromadb>=0.5.5
langchain-core>=0.3.0
langchain-community>=0.3.0
httpx>=0.27.0
pytest>=8.3.0
PyPDF2>=3.0.1
python-docx>=1.1.2
Pillow>=10.4.0
pytesseract>=0.3.10
PyMuPDF>=1.24.9
